feat: add user detail page
